Implemented Collections with tests
The following were implemented and have tests:
* MonitorCollection
* PrognosticCollection
* DiagnosticCollection

These were moved from "collections.py" to "base_components.py" as they depend
heavily on the code of the components they contain, and it should be more
convenient to be able to look at all this code in one place.

The "mock" package was added to the dev requirements, as it is required to
run the tests for these objects.