Added consistency base test for components, and applied to HeldSuarez
The new ComponentBase class in test_components will compare
component output to a cached result, and pass if the output is the same.
If no cache is present, the test will fail and the current result will be cached.

Currently xarray is used for caching, so these tests must only have DataArray
values in the outputs.

Other changes:
* HeldSuarez was added to climt namespace
* Cached output was added for HeldSuarez