Added NetCDFMonitor
In doing so, also corrected the labelling of the dimensions for tau
from the Frierson06LongwaveRadiation, and changed levels
from half/full to mid/interface.
DependencyException and IOException were also made in the making
of this object. The object only works if netcdf4 is installed, if not it
raises a DependencyException.