Generate random numbers for each radiation call
When using McICA, generate random numbers (determining which
wavelengths see cloud and which do not) each time the radiation
is called and heating rates are calculated. This means that for
a state that does not change, calling the radiation multiple times
gives different fluxes and heating rates each time, as different
wavelengths are subject to different conditions (cloud or no
cloud). Averaging over multiple calls is a better approximation
of the true atmosphere, where all wavelengths see some cloud.